+++ src/main/java/snw/rfm/api/events/HunterCatchPlayerEvent.java
@@ -11,6 +11,7 @@
 package snw.rfm.api.events;
 
 import org.bukkit.entity.Player;
+import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
 import org.jetbrains.annotations.NotNull;
@@ -22,11 +23,12 @@ import org.jetbrains.annotations.NotNull;
  * <p>
  * 此事件完成后，玩家被淘汰。
  */
-public final class HunterCatchPlayerEvent extends Event {
+public final class HunterCatchPlayerEvent extends Event implements Cancellable {
     private static final HandlerList handlers = new HandlerList();
     private final Player whoBeCatched;
     private final Player catcher;
     private final int playerRemaining;
+    private boolean isCancelled = false;
 
     public HunterCatchPlayerEvent(Player catched, Player catcher, int playerRemaining) {
         this.whoBeCatched = catched;
@@ -69,5 +71,15 @@ public final class HunterCatchPlayerEvent extends Event {
     @Override
     public HandlerList getHandlers() {
         return handlers;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return isCancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean cancel) {
+        isCancelled = cancel;
     }
 }